Related to #26082

This is only about making blending `BilevelImage`s faster. Taking it from ~20% of the profile to ~2% :^)

From:
```
Benchmark 1: BuildLagom/bin/pdf --render out.png PDF/Mellor_ACTITC_10.pdf --page 4
  Time (mean ± σ):     630.9 ms ±   9.7 ms    [User: 582.8 ms, System: 47.7 ms]
  Range (min … max):   622.1 ms … 655.9 ms    10 runs
```

To:
```
Benchmark 1: BuildLagom/bin/pdf --render out.png PDF/Mellor_ACTITC_10.pdf --page 4
  Time (mean ± σ):     521.2 ms ±   4.0 ms    [User: 475.4 ms, System: 45.5 ms]
  Range (min … max):   516.3 ms … 527.9 ms    10 runs
```