@nico 2025-10-03
> I think the real speedup for this code is to not do one bool per pixel but pack them into bits

Oh, looks like it's that way already actually! So maybe that isn't even that hard to do now that the code is in BilevelImage. (But for a future PR.)