@LucasChollet 2025-10-03
> I think the real speedup for this code is to not do one bool per pixel but pack them into bits, and then do 64 pixels in parallel. But if this enough to basically remove it from the profile for now, this is good enough for now :)

I even though of using SIMD for that, but most of the input images are actually very small ~20x~20 px. So I guess that "only" 8 bits is a good tradeoff.