@nico 2025-10-05
Does the last commit actually do anything? We always call `combine()` with a bool arg, no?