@nico 2025-10-06
-> #26248, thanks!

That does look like I saw it already at some point 😅